# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for Jobberwocky API",
        "contact": {
            "name": "Matias Lancry",
            "email": "lancrymatias@live.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subscriptions": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Create a new subscription",
                "description": "Creates a new email subscription with unique email validation.",
                "operationId": "storeSubscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email address for the subscription",
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created successfully"
                    },
                    "400": {
                        "description": "Bad Request: Email is missing or invalid format, or email already exists"
                    },
                    "500": {
                        "description": "Internal Server Error: An error occurred while processing the request"
                    }
                }
            }
        }
    },
...
}